# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EventSphere API - College Events Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: dateRange
          in: query
          schema:
            type: string
            enum: [any, today, this_week, this_month]
        - name: format
          in: query
          schema:
            type: string
            enum: [in_person, online]
        - name: department
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, ongoing, completed, waitlist]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 9
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{eventId}/approve:
    post:
      operationId: approveEvent
      tags: [events]
      summary: Approve or reject an event (admin)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveEventBody"
      responses:
        "200":
          description: Event approval status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Platform analytics overview
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/trends:
    get:
      operationId: getAnalyticsTrends
      tags: [analytics]
      summary: Event participation trends
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 14d, 30d]
            default: 14d
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendDataPoint"
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        department:
          type: string
        venue:
          type: string
        format:
          type: string
          enum: [in_person, online]
        startDate:
          type: string
        endDate:
          type: string
        capacity:
          type: integer
        registeredCount:
          type: integer
        status:
          type: string
          enum: [draft, pending, approved, rejected, ongoing, completed]
        isFeatured:
          type: boolean
        price:
          type: number
        imageUrl:
          type: string
        organizerId:
          type: integer
        organizerName:
          type: string
        organizerDept:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - venue
        - format
        - startDate
        - capacity
        - registeredCount
        - status
        - isFeatured
        - price
        - organizerId
        - organizerName
        - organizerDept
        - createdAt
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            mediaGallery:
              type: array
              items:
                type: string
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            averageRating:
              type: number
            reviewCount:
              type: integer
          required: [mediaGallery, reviews, averageRating, reviewCount]
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [events, total, page, totalPages]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        department:
          type: string
        venue:
          type: string
        format:
          type: string
          enum: [in_person, online]
        startDate:
          type: string
        endDate:
          type: string
        capacity:
          type: integer
        price:
          type: number
        imageUrl:
          type: string
      required: [title, category, venue, format, startDate, capacity]
    ApproveEventBody:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject]
        reason:
          type: string
      required: [action]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        avatarUrl:
          type: string
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string
      required: [id, userId, userName, rating, comment, createdAt]
...
    AnalyticsOverview:
      type: object
      properties:
        totalRegistrations:
          type: integer
        registrationsChange:
          type: number
        averageAttendanceRate:
          type: number
        attendanceChange:
          type: number
        overallFeedbackScore:
          type: number
        totalReviews:
          type: integer
      required:
        - totalRegistrations
        - registrationsChange
        - averageAttendanceRate
        - attendanceChange
        - overallFeedbackScore
        - totalReviews
    TrendDataPoint:
      type: object
      properties:
        date:
          type: string
        registrations:
          type: integer
        attendance:
          type: integer
      required: [date, registrations, attendance]
...